feat: ios: basic modals for 3 remaining errors
