+++ ios/NativeSigner/Modals/Errors/ErrorBottomModalViewModel.swift
@@ -92,55 +92,15 @@ struct ErrorBottomModalViewModel {
         )
     }
 
-    static func signingInvalidNetworkVersion(
-        _ networkName: String,
+    static func transactionSigningError(
+        message: String,
         _ action: @escaping @autoclosure () -> Void = {}()
     ) -> ErrorBottomModalViewModel {
         ErrorBottomModalViewModel(
-            title: Localizable.TransactionSign.Error.InvalidNetworkVersion.title(networkName),
-            content: Localizable.TransactionSign.Error.InvalidNetworkVersion.message(networkName),
-            steps: [
-                .init(
-                    step: "1",
-                    content: Localizable.signingInvalidNetworkVersionStepOne()
-                ),
-                .init(
-                    step: "2",
-                    content: AttributedString(Localizable.TransactionSign.Error.InvalidNetworkVersion.step2.string)
-                ),
-                .init(
-                    step: "3",
-                    content: AttributedString(Localizable.TransactionSign.Error.InvalidNetworkVersion.step3.string)
-                )
-            ],
-            secondaryAction: .init(label: Localizable.TransactionSign.Action.error.key, action: action)
-        )
-    }
-
-    static func signingUnknownNetwork(_ action: @escaping @autoclosure () -> Void = {}())
-        -> ErrorBottomModalViewModel {
-        ErrorBottomModalViewModel(
-            title: Localizable.TransactionSign.Error.UnknownNetwork.title.string,
-            content: Localizable.TransactionSign.Error.UnknownNetwork.message.string,
-            steps: [
-                .init(
-                    step: "1",
-                    content: Localizable.signingUnknownNetworkStepOne()
-                ),
-                .init(
-                    step: "2",
-                    content: AttributedString(Localizable.TransactionSign.Error.UnknownNetwork.step2.string)
-                ),
-                .init(
-                    step: "3",
-                    content: AttributedString(Localizable.TransactionSign.Error.UnknownNetwork.step3.string)
-                ),
-                .init(
-                    step: "4",
-                    content: AttributedString(Localizable.TransactionSign.Error.UnknownNetwork.step4.string)
-                )
-            ],
-            secondaryAction: .init(label: Localizable.TransactionSign.Action.error.key, action: action)
+            title: Localizable.TransactionSign.Error.Generic.title.string,
+            content: Localizable.GenericErrorModal.Label.messagePrefix.string,
+            details: message,
+            secondaryAction: .init(label: Localizable.GenericErrorModal.Action.ok.key, action: action)
         )
     }
 
@@ -172,18 +132,6 @@ struct ErrorBottomModalViewModel {
     ) -> ErrorBottomModalViewModel {
         ErrorBottomModalViewModel(
             title: Localizable.GenericErrorModal.Label.title.string,
-            content: Localizable.GenericErrorModal.Label.messagePrefix.string,
-            details: message,
-            secondaryAction: .init(label: Localizable.GenericErrorModal.Action.ok.key, action: action)
-        )
-    }
-
-    static func transactionSigningError(
-        message: String,
-        _ action: @escaping @autoclosure () -> Void = {}()
-    ) -> ErrorBottomModalViewModel {
-        ErrorBottomModalViewModel(
-            title: Localizable.TransactionSign.Error.Generic.title.string,
             content: Localizable.GenericErrorModal.Label.messagePrefix.string,
             details: message,
             secondaryAction: .init(label: Localizable.GenericErrorModal.Action.ok.key, action: action)
