+++ ios/NativeSigner/Screens/Scan/Password/EnterPasswordModal.swift
